requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
jinja2
email-validator
python-dotenv
